extract stream_chunk
