+++ massgen/backend/response.py
@@ -15,9 +15,9 @@ from openai import AsyncOpenAI
 from ..api_params_handler import ResponseAPIParamsHandler
 from ..formatter import ResponseFormatter
 from ..logger_config import log_backend_agent_message, log_stream_chunk, logger
+from ..stream_chunk import ChunkType, TextStreamChunk
 from .base import FilesystemSupport, StreamChunk
 from .base_with_mcp import MCPBackend
-from .stream_chunk import ChunkType, TextStreamChunk
 
 
 class ResponseBackend(MCPBackend):
