+++ massgen/stream_chunk/base.py
