+++ massgen/stream_chunk/multimodal.py
