+++ massgen/stream_chunk/text.py
